FinalYükleme/static/requirements.txt
Flask
flask-mysqldb
python-dotenv
